Add GetScreenRay function to create a ray from a screen position.

+++ Sources/Modeler/Camera.cpp
@@ -4,14 +4,14 @@
 namespace Modeler
 {
 
-static const float MinZoomDistance = 1.0f;
+static const double MinZoomDistance = 1.0f;
 
-Camera::Camera (const glm::vec3& eye,
-				const glm::vec3& center,
-				const glm::vec3& up,
-				float fieldOfView,
-				float nearPlane,
-				float farPlane) :
+Camera::Camera (const glm::dvec3& eye,
+				const glm::dvec3& center,
+				const glm::dvec3& up,
+				double fieldOfView,
+				double nearPlane,
+				double farPlane) :
 	eye (eye),
 	center (center),
 	up (up),
@@ -22,17 +22,17 @@ Camera::Camera (const glm::vec3& eye,
 	
 }
 
-const glm::vec3& Camera::GetEye () const
+const glm::dvec3& Camera::GetEye () const
 {
 	return eye;
 }
 
-const glm::vec3& Camera::GetCenter () const
+const glm::dvec3& Camera::GetCenter () const
 {
 	return center;
 }
 
-const glm::vec3& Camera::GetUp () const
+const glm::dvec3& Camera::GetUp () const
 {
 	return up;
 }
@@ -44,31 +44,31 @@ glm::mat4 Camera::GetViewMatrix () const
 
 glm::mat4 Camera::GetProjectionMatrix (int width, int height) const
 {
-	return glm::perspective (glm::radians (fieldOfViewY), (float) width / (float) height, nearPlane, farPlane);
+	return glm::perspective (glm::radians (fieldOfViewY), (double) width / (double) height, nearPlane, farPlane);
 }
 
-static glm::vec3 Rotate (const glm::vec3& vec, const glm::vec3& center, float angle, const glm::vec3& normal)
+static glm::dvec3 Rotate (const glm::dvec3& vec, const glm::dvec3& center, double angle, const glm::dvec3& normal)
 {
-	glm::vec3 result = vec - center;
+	glm::dvec3 result = vec - center;
 	result = glm::rotate (result, angle, normal);
-	return glm::vec3 (result + center);
+	return glm::dvec3 (result + center);
 }
 
-void Camera::Orbit (float angleX, float angleY)
+void Camera::Orbit (double angleX, double angleY)
 {
-	float radAngleX = glm::radians (angleX);
-	float radAngleY = glm::radians (angleY);
+	double radAngleX = glm::radians (angleX);
+	double radAngleY = glm::radians (angleY);
 	
-	glm::vec3 viewDirection = glm::normalize (center - eye);
-	glm::vec3 upDirection = glm::normalize (up);
-	glm::vec3 horizontalDirection = glm::normalize (glm::cross (viewDirection, upDirection));
+	glm::dvec3 viewDirection = glm::normalize (center - eye);
+	glm::dvec3 upDirection = glm::normalize (up);
+	glm::dvec3 horizontalDirection = glm::normalize (glm::cross (viewDirection, upDirection));
 
 	// TODO: different center
-	glm::vec3 orbitCenter = center;
+	glm::dvec3 orbitCenter = center;
 
-	float originalAngle = glm::angle (glm::normalize (eye - center), upDirection);
-	float newAngle = originalAngle + radAngleY;
-	if (Geometry::IsGreater (newAngle, 0.0f) && Geometry::IsLower (newAngle, glm::pi<float> ())) {
+	double originalAngle = glm::angle (glm::normalize (eye - center), upDirection);
+	double newAngle = originalAngle + radAngleY;
+	if (Geometry::IsGreater (newAngle, 0.0) && Geometry::IsLower (newAngle, glm::pi<double> ())) {
 		eye = Rotate (eye, orbitCenter, radAngleY, horizontalDirection);
 		center = Rotate (center, orbitCenter, radAngleY, horizontalDirection);
 	}
@@ -77,26 +77,26 @@ void Camera::Orbit (float angleX, float angleY)
 	center = Rotate (center, orbitCenter, radAngleX, upDirection);
 }
 
-void Camera::Pan (float diffX, float diffY)
+void Camera::Pan (double diffX, double diffY)
 {
-	glm::vec3 viewDirection = glm::normalize (center - eye);
-	glm::vec3 upDirection = glm::normalize (up);
-	glm::vec3 horizontalDirection = glm::normalize (glm::cross (viewDirection, upDirection));
-	glm::vec3 verticalDirection = glm::normalize (glm::cross (horizontalDirection, viewDirection));
+	glm::dvec3 viewDirection = glm::normalize (center - eye);
+	glm::dvec3 upDirection = glm::normalize (up);
+	glm::dvec3 horizontalDirection = glm::normalize (glm::cross (viewDirection, upDirection));
+	glm::dvec3 verticalDirection = glm::normalize (glm::cross (horizontalDirection, viewDirection));
 
-	float eyeCenterDistance = glm::length (center - eye);
-	float relDiffX = diffX / 1000.0f * eyeCenterDistance;
-	float relDiffY = diffY / 1000.0f * eyeCenterDistance;
+	double eyeCenterDistance = glm::length (center - eye);
+	double relDiffX = diffX / 1000.0f * eyeCenterDistance;
+	double relDiffY = diffY / 1000.0f * eyeCenterDistance;
 
 	eye = eye + horizontalDirection * relDiffX + verticalDirection * relDiffY;
 	center = center + horizontalDirection * relDiffX + verticalDirection * relDiffY;
 }
 
-void Camera::Zoom (float ratio)
+void Camera::Zoom (double ratio)
 {
-	glm::vec3 direction = center - eye;
-	float distance = glm::length (direction);
-	float movement = distance * ratio;
+	glm::dvec3 direction = center - eye;
+	double distance = glm::length (direction);
+	double movement = distance * ratio;
 	bool zoomIn = ratio > 0.0f;
 	if (zoomIn && distance < MinZoomDistance) {
 		return;
@@ -104,22 +104,22 @@ void Camera::Zoom (float ratio)
 	eye = eye + glm::normalize (direction) * movement;
 }
 
-void Camera::ZoomToSphere (const glm::vec3& sphereCenter, float sphereRadius, int width, int height)
+void Camera::ZoomToSphere (const glm::dvec3& sphereCenter, double sphereRadius, int width, int height)
 {
-	glm::vec3 offetToOrigin = center - sphereCenter;
+	glm::dvec3 offetToOrigin = center - sphereCenter;
 	center = sphereCenter;
 	eye = eye - offetToOrigin;
 	
-	float calcFieldOfView = 0.0;
-	float widthHeightRatio = (float) width / (float) height;
+	double calcFieldOfView = 0.0;
+	double widthHeightRatio = (double) width / (double) height;
 	if (widthHeightRatio >= 1.0f) {
-		calcFieldOfView = fieldOfViewY / 2.0f;
+		calcFieldOfView = fieldOfViewY / 2.0;
 	} else {
-		calcFieldOfView = fieldOfViewY / 2.0f * widthHeightRatio;
+		calcFieldOfView = fieldOfViewY / 2.0 * widthHeightRatio;
 	}
 	
-	float distance = sphereRadius / glm::sin (glm::radians (calcFieldOfView));
-	glm::vec3 centerToEyeDirection = glm::normalize (eye - center);
+	double distance = sphereRadius / glm::sin (glm::radians (calcFieldOfView));
+	glm::dvec3 centerToEyeDirection = glm::normalize (eye - center);
 	eye = center + centerToEyeDirection * distance;
 
 	// TODO: different center
